+++ sdk/identity/Azure.Identity/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
  - Allow `VisualStudioCredential` on non-Windows platforms
  - Added `AzureDeveloperCliCredential` for Azure Developer CLI
+ - Added `Logout` and `LogoutAsync` methods to supporting credentials
 
 ### Breaking Changes
 - Previously, if environment variables for username and password auth are set in addition to the AZURE_CLIENT_CERTIFICATE_PATH, EnvironmentCredential would select the `UsernamePasswordCredential`. After this change, `ClientCertificateCredential` will be selected, which is consistent with all other languages. This is potentially a behavioral breaking change.